MID-2936: keep services at model package level
