+++ mid-java-client-core/src/main/java/ch/swisscom/mid/client/model/GeofencingAdditionalServiceResponse.java
@@ -16,11 +16,27 @@
  *
  */
 
-package ch.swisscom.mid.client.model.service;
+/*
+ *
+ *  * Copyright 2021-2026 Swisscom (Schweiz) AG
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+package ch.swisscom.mid.client.model;
 
 import ch.swisscom.mid.client.config.DefaultConfiguration;
-import ch.swisscom.mid.client.model.AdditionalServiceResponse;
-import ch.swisscom.mid.client.model.GeofencingErrorCode;
 
 /**
  * The information returned when the Geofencing additional service is requested to Mobile ID.
