+++ mid-java-client-core/src/main/java/ch/swisscom/mid/client/model/SignatureValidationAdditionalService.java
@@ -17,6 +17,24 @@
  */
 
 /*
+ *
+ *  * Copyright 2021-2026 Swisscom (Schweiz) AG
+ *  *
+ *  * Licensed under the Apache License, Version 2.0 (the "License");
+ *  * you may not use this file except in compliance with the License.
+ *  * You may obtain a copy of the License at
+ *  *
+ *  * http://www.apache.org/licenses/LICENSE-2.0
+ *  *
+ *  * Unless required by applicable law or agreed to in writing, software
+ *  * distributed under the License is distributed on an "AS IS" BASIS,
+ *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  * See the License for the specific language governing permissions and
+ *  * limitations under the License.
+ *
+ */
+
+/*
  * Copyright 2021 Swisscom (Schweiz) AG
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -31,22 +49,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ch.swisscom.mid.client.model.service;
+package ch.swisscom.mid.client.model;
 
 import ch.swisscom.mid.client.config.DefaultConfiguration;
-import ch.swisscom.mid.client.model.UserLanguage;
 
-public class UserLangAdditionalService extends AdditionalService {
+public class SignatureValidationAdditionalService extends AdditionalService {
 
-    private final UserLanguage userLanguage;
-
-    public UserLangAdditionalService(UserLanguage userLanguage) {
-        super(DefaultConfiguration.ADDITIONAL_SERVICE_USER_LANG_URI);
-        this.userLanguage = userLanguage;
-    }
-
-    public UserLanguage getUserLanguage() {
-        return userLanguage;
+    public SignatureValidationAdditionalService() {
+        super(DefaultConfiguration.ADDITIONAL_SERVICE_SIGNATURE_VALIDATION_URI);
     }
 
 }
